Add a Cypress test
