+++ .github/workflows/bashlib.sh
@@ -38,10 +38,10 @@ default-setup-command() {
 }
 
 apt-get-install() {
-    say "::group::apt-get install dependencies"
-    sudo apt-get update && sudo apt-get install --yes \
-      libsasl2-dev
-    say "::endgroup::"
+  say "::group::apt-get install dependencies"
+  sudo apt-get update && sudo apt-get install --yes \
+    libsasl2-dev
+  say "::endgroup::"
 }
 
 pip-upgrade() {
@@ -167,7 +167,8 @@ cypress-run() {
   if [[ -z $CYPRESS_KEY ]]; then
     $cypress --spec "cypress/integration/$page" --browser "$browser"
   else
-    export CYPRESS_RECORD_KEY=`echo $CYPRESS_KEY | base64 --decode`
+    CYPRESS_RECORD_KEY=$(echo "$CYPRESS_KEY" | base64 --decode)
+    export CYPRESS_RECORD_KEY
     # additional flags for Cypress dashboard recording
     $cypress --spec "cypress/integration/$page" --browser "$browser" \
       --record --group "$group" --tag "${GITHUB_REPOSITORY},${GITHUB_EVENT_NAME}" \
