+++ superset-frontend/cypress-base/cypress/integration/explore/link.test.ts
@@ -23,7 +23,7 @@
 import rison from 'rison';
 import shortid from 'shortid';
 import { interceptChart } from 'cypress/utils';
-import { HEALTH_POP_FORM_DATA_DEFAULTS } from './visualizations/shared.helper';
+import { HEALTH_POP_FORM_DATA_DEFAULTS } from './visualizations/constants';
 
 const apiURL = (endpoint: string, queryObject: Record<string, unknown>) =>
   `${endpoint}?q=${rison.encode(queryObject)}`;
